@@ -25,7 +25,7 @@
     "url": "https://github.com/VSCodeVim/Vim/issues"
   },
   "engines": {
-    "vscode": "^0.10.1"
+    "vscode": "^0.10.8"
   },
   "categories": [
     "Other"
@@ -144,23 +144,23 @@
     "vscode:prepublish": "node ./node_modules/vscode/bin/compile",
     "compile": "node ./node_modules/vscode/bin/compile -watch -p ./",
     "test": "node ./node_modules/vscode/bin/test",
-    "postinstall": "gulp init"
-  },
-  "devDependencies": {
-    "gulp": "^3.9.0",
-    "gulp-mocha": "^2.2.0",
-    "gulp-shell": "^0.5.1",
-    "gulp-soften": "^0.0.1",
-    "gulp-trimlines": "^1.0.0",
-    "gulp-tslint": "^3.6.0",
-    "gulp-typescript": "^2.9.2",
-    "gulp-typings": "^1.1.0",
-    "tslint": "^3.2.1",
-    "typescript": "^1.6.2",
-    "typings": "^0.6.8",
-    "vscode": "0.10.x"
+    "postinstall": "node ./node_modules/vscode/bin/install && gulp init"
   },
   "dependencies": {
-    "lodash": "^3.10.1"
+    "lodash": "^4.5.1"
+  },
+  "devDependencies": {
+    "gulp": "^3.9.1",
+    "gulp-mocha": "^2.2.0",
+    "gulp-shell": "^0.5.2",
+    "gulp-soften": "^0.0.1",
+    "gulp-trimlines": "^1.0.0",
+    "gulp-tslint": "^4.3.2",
+    "gulp-typescript": "^2.12.0",
+    "gulp-typings": "^1.1.0",
+    "tslint": "^3.5.0",
+    "typescript": "^1.8.2",
+    "typings": "^0.6.8",
+    "vscode": "^0.11.1"
   }
 }
